Port.

+++ client_main.c
@@ -4,7 +4,7 @@
 #include <process.h>
 #include <threads.h>
 
-#include "lib/bstrlib.h"
+#include "bstrlib.h"
 
 #include "debug.h"
 #include "client_protocol.h"
@@ -46,10 +46,10 @@ int main(int argc, char *argv[])
         return -1;
     }
 
-    int port = 0;
+    unsigned short port = 0;
     if (2 < argc)
     {
-        port = atoi(argv[2]);
+        port = (unsigned short) atoi(argv[2]);
         if (!port)
         {
             port = PORT;
